无人机物体箱定位系统（Flask端云）/drone_side/requirements.txt
torch>=1.12.0
torchvision>=0.13.0
opencv-python>=4.6.0
ultralytics>=8.0.0
pyzbar>=0.1.9
pymavlink>=2.4.37
requests>=2.28.0
numpy>=1.21.0
Pillow>=9.0.0
websockets>=10.0
pycryptodome>=3.20.0
sympy>=1.12
